battleship_testing.c
this branch will serve as a backup layer. The changes made by the team will be committed and pulled here at first before getting to the main so as to ensure that the main remains unharmed in case there is a miscommunication or unresolved major bug in the current version of the code.